[FIXED] pack p3d's alpha
Former-commit-id: f60b05086e958f33204c76254adcd34a01dd53fb

+++ easyrespacker/src/librespacker/PackParticle3D.h
@@ -18,7 +18,7 @@ public:
 		float angle, angle_var;
 
 		uint32_t col_mul, col_add;
-		float alpha_start, alpha_end;
+		uint8_t alpha_start, alpha_end;
 
 		const IPackNode* node;
 	};
